### 2023-04-02 #5326 @editorialbot
[PRE REVIEW]: predictNMB: An R package to estimate if or when a clinical prediction model is worthwhile
**Submitting author:** @RWParsons (<a href="http://orcid.org/0000-0002-6053-8174">Rex Parsons</a>)
**Repository:** https://github.com/ropensci/predictNMB
**Branch with paper.md** (empty if default branch): joss-paper
**Version:** v0.1.0
