### 2023-04-02 #5326 @Kevin-Mattheus-Moerman
This R package has been peer reviewed and accepted through ropensci (https://docs.ropensci.org/predictNMB/)
