+++ paper/paper.bib
@@ -6,7 +6,8 @@
   number={1},
   pages={1--7},
   year={2019},
-  publisher={BioMed Central}
+  publisher={BioMed Central},
+  doi={10.1186/s12916-019-1425-3}
 }
 
 @article{hendriksen2013diagnostic,
