### 2023-04-02 #5328 @RWParsons
Hi @Kevin-Mattheus-Moerman - thanks for being editor for this submission.

- I've just updated the paper to add the missing DOI - thanks for picking this one up.
- I hadn't made a ZENODO archive for the package so I have just made one: 10.5281/zenodo.7793463
- Yes, I confirm that the verion tag should be v0.1.0.
