### 2023-04-02 #5328 @RWParsons
@Kevin-Mattheus-Moerman No worries. I've just updated the license and authors of the zenodo entry. Sorry, I wasn't sure whether the zenodo should match the package's DESCRIPTION file or the paper.
