### 2023-04-03 ropensci/predictNMB#4 @RWParsons
Release predictNMB 0.1.0
First release of software predictNMB for submission. Used for submission to [zenodo ](https://doi.org/10.5281/zenodo.7793461) and associated submission to [JOSS](https://github.com/openjournals/joss-reviews/issues/5328).
