### 2023-04-03 #5326 @RWParsons
Hi @Kevin-Mattheus-Moerman,

I'm not sure whether this should be posted in this issue or the new one (https://github.com/openjournals/joss-reviews/issues/5328) so I'm sorry if this is in the wrong spot.

In the OP from the editorial bot it asks for suggested reviewers. I have asked Emi Tanaka ('emitanaka' on GitHub) who reviewed the software for ropensci and she said she would be happy to review the JOSS paper. From the list of JOSS reviewers, 'stulacy' and 'epiben' look to have relevant experience.
