### 2023-04-04 #5328 @Kevin-Mattheus-Moerman
@RWParsons this reference seems poorly formatted in the paper, can you check?:
https://journals.plos.org/plosmedicine/article?id=10.1371/journal.pmed.1001381

Also, I think you can include this DOI for it?
[10.1371/journal.pmed.1001381](https://doi.org/10.1371/journal.pmed.1001381)
