+++ paper/paper.bib
@@ -17,17 +17,19 @@
   volume={11},
   pages={129--141},
   year={2013},
+  doi={10.1111/jth.12262},
   publisher={Wiley Online Library}
 }
 
 @article{steyerberg2013prognosis,
-  title={Prognosis Research Strategy (PROGRESS) 3: prognostic model research},
-  author={Steyerberg, Ewout W and Moons, Karel GM and van der Windt, Danielle A and Hayden, Jill A and Perel, Pablo and Schroter, Sara and Riley, Richard D and Hemingway, Harry and Altman, Douglas G and PROGRESS Group},
+  title={Prognosis Research Strategy (PROGRESS) 3: Prognostic Model Research},
+  author={Steyerberg, Ewout W and Moons, Karel GM and van der Windt, Danielle A and Hayden, Jill A and Perel, Pablo and Schroter, Sara and Riley, Richard D and Hemingway, Harry and Altman, Douglas G},
   journal={PLoS medicine},
   volume={10},
   number={2},
   pages={e1001381},
   year={2013},
+  doi={10.1371/journal.pmed.1001381},
   publisher={Public Library of Science San Francisco, USA}
 }
 
@@ -39,6 +41,7 @@
   number={3},
   pages={201--209},
   year={2006},
+  doi={10.7326/0003-4819-144-3-200602070-00009},
   publisher={American College of Physicians}
 }
 
