### 2023-04-05 #5326 @Kevin-Mattheus-Moerman
@RWParsons if you had your software reviewed and accepted through ropensci then we use a streamlined process. So, as editor and the only one at JOSS who reviews your paper and helps process final steps. We do not re-do a formal software review again, see also our editorial guide on this: https://joss.readthedocs.io/en/latest/editing.html#processing-of-ropensci-reviewed-or-pyopensci-reviewed-and-accepted-submissions
